+++ src/api/v1/auth/model.py
