+++ src/api/v1/users/images/model.py
