+++ src/api/v1/users/images/service.py
@@ -1,4 +1,4 @@
-from src.api.users.images.model import Image
+from src.api.v1.users.images.model import Image
 from mongoengine.queryset.visitor import Q
